requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
pyyaml>=6,<7
